+++ create-ao-dapp/src/utils/aoform.ts
@@ -1,9 +1,6 @@
 import { writeFile } from "node:fs/promises";
 import { Document } from "yaml";
-import { $ } from "zx";
 import type { PackageManager } from "../types";
-
-const AOFORM_PACKAGE = "github:Autonomous-Finance/aoform#4cbb877ca646e8823fd7ce4fdae0f9c38aef4830";
 
 export default class AoFormGenerator {
   projectPath: string;
@@ -14,14 +11,14 @@ export default class AoFormGenerator {
     projectPath,
     processName,
     packageManager,
-  }: { projectPath: string; processName: string; packageManager: PackageManager }) {
+  }: {
+    projectPath: string;
+    processName: string;
+    packageManager: PackageManager;
+  }) {
     this.projectPath = projectPath;
     this.processName = processName;
     this.packageManager = packageManager;
-  }
-
-  async installAoForm() {
-    return await $`cd ${this.projectPath} && ${this.packageManager} install --save-dev ${AOFORM_PACKAGE}`;
   }
 
   async generateAoFormYaml() {
@@ -39,17 +36,20 @@ export default class AoFormGenerator {
           },
           {
             name: "Powered-By",
-            value: "Ao Starter Kit",
-          }
+            value: "create-ao-dapp",
+          },
         ],
       },
     ]);
 
     doc.commentBefore = `# AoForm Configuration
 # This file is used to configure the AoForm package
-# For more information, visit`;
+# For more information, visit https://github.com/Autonomous-Finance/aoform`;
 
-    return await writeFile(`${this.projectPath}/ao/${this.processName}/aoform.yaml`, String(doc));
+    return await writeFile(
+      `${this.projectPath}/ao/${this.processName}/aoform.yaml`,
+      String(doc)
+    );
   }
 
   async addDeployScript() {
@@ -57,6 +57,9 @@ export default class AoFormGenerator {
 	
 aoform apply -f ./ao/${this.processName}/aoform.yaml`;
 
-    await writeFile(`${this.projectPath}/ao/${this.processName}/scripts/deploy.sh`, deployScript);
+    await writeFile(
+      `${this.projectPath}/ao/${this.processName}/scripts/deploy.sh`,
+      deployScript
+    );
   }
 }