+++ create-ao-dapp/templates/ao/lua/base/src/reset_modules.lua
@@ -1,0 +1,1 @@
+package.loaded["lib.process_lib"] = nil