+++ create-ao-dapp/templates/ao/lua/squishy/src/reset_modules.lua
@@ -1,1 +1,0 @@
-package.loaded["process_lib"] = nil