+++ create-ao-dapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-ao-dapp
 
+## 0.1.1
+
+### Patch Changes
+
+- update permaweb deployment script and docs
+
+## 0.1.0
+
+### Minor Changes
+
+- implement amalg.lua and permaweb deploy
+
 ## 0.0.2
 
 ### Patch Changes